Refactor concurrency limiter classes
